should switch services on error while in auto mode and clean up the filess
