+++ binnacle/lib/services/compass_service.dart
@@ -3,6 +3,7 @@ import '../models/compass_model.dart';
 import '../models/compass_service_interface.dart';
 import 'package:flutter_compass/flutter_compass.dart';
 import '../models/service_data.dart';
+import './service_wrapper_interface.dart';
 
 class CompassService extends ICompassService {
   
@@ -11,19 +12,27 @@ class CompassService extends ICompassService {
 
 
   CompassService() {
-    print('Initializing Compass Service');
     _subscription = FlutterCompass.events.listen((data) => _compassStream.sink.add(CompassModel(direction: data)));
-    // _compassStream.addStream(FlutterCompass.events.map((double d) => new CompassModel(direction: d)));
-    //; // Do I need to listen here?
+
   }
   dispose() async {
     await _subscription.pause();
     
     await _compassStream.close();
 
-    // await _compassStream.close();
-
   }
   StreamController<CompassModel> get compassStream => _compassStream;
+
+}
+
+class CompassServiceWrapper implements ServiceWrapper{
+  final ServiceData _serviceData = ServiceData('compass', 'flutter compass', 1);
+  final bool _default = true;
+
+  CompassServiceWrapper();
+
+  get service =>  CompassService();
+  ServiceData get serviceData => this._serviceData;
+  bool get isDefault => this._default;
 
 }
