+++ binnacle/lib/services/test_compass_service.dart
@@ -1,6 +1,8 @@
 import 'dart:async';
 import '../models/compass_model.dart';
 import '../models/compass_service_interface.dart';
+import './service_wrapper_interface.dart';
+import '../models/service_data.dart';
 class TestCompassService extends ICompassService {
 
   StreamController<CompassModel> _compassStream = StreamController();
@@ -21,4 +23,15 @@ class TestCompassService extends ICompassService {
     await _compassStream.close();
   }
   StreamController<CompassModel> get compassStream => _compassStream;
+}
+
+class MockCompassServiceWrapper extends ServiceWrapper{
+  final ServiceData _serviceData = ServiceData('compass', 'mock compass', 0);
+  final bool _default = false;
+
+  MockCompassServiceWrapper();
+
+  get service =>  TestCompassService();
+  ServiceData get serviceData => this._serviceData;
+  bool get isDefault => this._default;
 }
